Move main and compliance.go, rename compliance package
