+++ compliance/collection/compliance_checks/compliance_checks_bench_test.go
@@ -1,4 +1,4 @@
-package compliance
+package compliance_checks
 
 import (
 	"encoding/json"
